refactor: moving Drain into drain.rs
